Make master compile on the mac.

+++ src/ee/storage/ElasticContext.cpp
@@ -67,7 +67,10 @@ TableStreamerContext* ElasticContext::cloneForTruncatedTable(PersistentTableSurg
             continue;
         } else {
             char errMsg[1024];
-            snprintf(errMsg, 1024, "Received an unrecognized return value %jd from handleStreamMore()", retCode);
+            snprintf(errMsg,
+                     1024,
+                     "Received an unrecognized return value %jd from handleStreamMore()",
+                     (intmax_t)retCode);
             LogManager::getThreadLogger(LOGGERID_HOST)->log(LOGLEVEL_ERROR, errMsg);
             break;
         }
